requirements.txt
agate==1.6.3
appdirs==1.4.4
attrs==23.1.0
Babel==2.12.1
black==23.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
db-dtypes==1.1.1
dbt-bigquery==1.5.3
dbt-core==1.5.2
dbt-extractor==0.4.1
diff-cover==7.6.1
exceptiongroup==1.1.2
future==0.18.3
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.3
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-dataproc==5.4.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
hologram==0.0.16
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.1.3
mashumaro==3.6
minimal-snowplow-tracker==0.0.2
msgpack==1.0.5
mypy-extensions==1.0.0
networkx==2.8.8
numpy==1.25.1
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pandas-gbq==0.19.2
parsedatetime==2.4
pathspec==0.11.1
platformdirs==3.8.1
pluggy==1.2.0
proto-plus==1.22.3
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydata-google-auth==1.8.1
Pygments==2.15.1
pytest==7.4.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.8.10
rsa==4.9
six==1.16.0
sqlfluff==2.1.2
sqlparse==0.4.3
tblib==2.0.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6